% Generated by roxygen2 (4.0.0): do not edit by hand
\name{gp_fit}
\alias{gp_fit}
\title{gp_fit fits a Gaussian process to the observed data}
\usage{
gp_fit(obs, X, pars = c(sigma_n = 1, tau = 1, l = 1), method = c("direct",
  "sequential", "kernlab", "cholesky", "conditional"), fit = TRUE,
  out_var = NULL, ...)
}
\arguments{
\item{obs}{a data frame of observations with columns obs$x and obs$y}

\item{X}{the desired points over which to predict}

\item{pars}{a named numeric specifying "sigma_n" for the (additive) noise and "l" for the covariance length scale}

\item{method}{select the method to use}

\item{out_var}{optionally force kernlab method to generate Cf using a different variance than sigma_n^2 assumed in the fit}

\item{fit}{logical, argument for the kernlab::gausspr method only}

\item{...}{potential additional arguments, unimplemented}
}
\value{
a list with items "mu", the expected Y values at X (mean of the posterior Gaussian process),
Sigma, the covariance matrix for the posterior Gaussian process, and
"loglik", the log likelihood of observering the given data under the process, marginalized over the prior
}
\description{
API not yet stable! Arguments to this function will probably be updated to be much more generic
}
\details{
so far treates the prior as mean 0 and covariance given by cov(X)
}

