% Generated by roxygen2 (4.0.0): do not edit by hand
\name{gp_mcmc}
\alias{gp_mcmc}
\title{Basic regression in Gaussian processes}
\usage{
gp_mcmc(x, y, init_pars = c(l = 1, sigma.n = 1), n = 10000, d.p = c(5, 5),
  s2.p = c(5, 5))
}
\arguments{
\item{x}{Observed x values, (vector or matrix with columns for each dimension of data)}

\item{y}{Vector of observed y values in the training data}

\item{init_pars}{the initial guesses for lengthscale l and process noise sigma_n}

\item{n}{iterations of the metropolis algorithm}
}
\value{
the MCMC output, as constructed by \link{metrop} from the mcmc package.
}
\description{
Basic regression in Gaussian processes
}
\details{
Currently assumes the covariance function.  By default we will use
the squared exponential (also called radial basis or Gaussian,
though it is not this that gives Gaussian process it's name;
any covariance function would do) as the covariance function,
\deqn{\operatorname{cov}(X_i, X_j) = \exp\left(-\frac{(X_i - X_j)^2}{2 \ell ^ 2}\right)}{}
}

