% Generated by roxygen2 (4.0.0): do not edit by hand
\name{gp_plot}
\alias{gp_plot}
\title{plot the gaussian process, true model, and fitted parametric model(s)}
\usage{
gp_plot(gp, f, p, f_est, p_est, f_alt, p_alt, x_grid, obs, seed)
}
\description{
plot the gaussian process, true model, and fitted parametric model(s)
}

