% Generated by roxygen2 (4.0.0): do not edit by hand
\name{posteriors_plot}
\alias{posteriors_plot}
\title{plot the gp posteriors}
\usage{
posteriors_plot(gp, priors)
}
\arguments{
\item{gp}{a gp fit}

\item{priors}{prior distribution functions so that they can be shown on plot as well}
}
\value{
the posteriors plot as histogram, with priors overlaid
}
\description{
plot the gp posteriors
}

