\name{profits_stats}
\alias{profits_stats}
\title{function to provide summary stats or the profits}
\usage{
  profits_stats(dt)
}
\arguments{
  \item{dt}{the data table from simulate_opt}
}
\value{
  a table of mean and variance in profit over the
  replicates, by method
}
\description{
  function to provide summary stats or the profits
}

