\name{sim_plots}
\alias{sim_plots}
\title{helper function to draw the plots of fish stock and harvest dynamics}
\usage{
  sim_plots(dt, seed = 1, harvest = FALSE)
}
\arguments{
  \item{dt}{the data.table from simulate_opt}
}
\value{
  plots of stock and harvest dynamics over time and
  replicates
}
\description{
  helper function to draw the plots of fish stock and
  harvest dynamics
}

