% Generated by roxygen2 (4.0.0): do not edit by hand
\name{simulate_opt}
\alias{simulate_opt}
\title{helper function to simulate the optimal policies of each model}
\usage{
simulate_opt(OPT, f, p, x_grid, h_grid, x0, z_g, profit, OptTime)
}
\description{
helper function to simulate the optimal policies of each model
}

