\name{optimal_policy}
\alias{optimal_policy}
\title{helper function to determine the optimal policies of each model}
\usage{
  optimal_policy(gp, f, f_alt, p, p_alt, x_grid, h_grid,
    sigma_g, sigma_g_alt, delta, xT, profit, reward,
    OptTime)
}
\description{
  helper function to determine the optimal policies of each
  model
}

