% Generated by roxygen2 (4.0.0): do not edit by hand
\name{par_F}
\alias{par_F}
\title{calculate transition matrix based on parameterized function}
\usage{
par_F(h, f, p, x_grid, sigma_g)
}
\arguments{
\item{h}{the harvest over which to compute.}

\item{f}{the transtion function, a function of state x, harvest h, and parameters p}

\item{p}{the parameters p of the transition function}

\item{x_grid}{the grid over which to solve}

\item{sigma_g}{the scale of the noise}
}
\value{
a list of matrices corresponding to the state transition matrix at each harvest level h
}
\description{
calculate transition matrix based on parameterized function
}

