% Generated by roxygen2 (4.0.0): do not edit by hand
\name{plot.gpfit}
\alias{plot.gpfit}
\title{Plot a Gaussian Process}
\usage{
\method{plot}{gpfit}(x, true, ...)
}
\arguments{
\item{x}{a gpfit object from \code{\link{gp_fit}}}

\item{true}{an optional data.frame with columns "x" and "y" containing target curve.}

\item{...}{additional arguments (not implemented)}
}
\value{
a ggplot object of the plot
}
\description{
Plot the mean (line) and 2 standard deviations (ribbon)
}

